Part 11: Access Control

+++ app/Http/Controllers/PostsController.php
@@ -7,11 +7,24 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+
+     
     public function index()
     {
         $posts = Post::orderBy('created_at', 'desc')->get();
@@ -71,6 +84,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthourized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -105,6 +123,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthourized Page');
+        }
+        
+        return view('posts.edit')->with('post', $post);
         $post->delete();
         return redirect('/posts')->with('success', 'Post Updated');
     }
